move SyntaxNode casting functions into ast.rs
